// src/app/core/models/studentTestAnswerSave.ts
import { EntityBase } from "./entityBase";


export interface StudentTestAnswerSave extends EntityBase {
    id: number;
    studentId?: number;
    testId : number; 
    accomodation: boolean | null;
    offline: boolean | null;
    fullScreenClosed: boolean | null;
    keyPress :boolean | null;
    leftExamArea: boolean | null;
    timeRemaining: string | null;
    answerText?: string | null;
    fileName: string | null;
}